OEA_Portal/OEA_Portal/core/services/requirements.txt
opencensus-ext-azure >= 1.0.8
Faker >= 8.0